// src/app/types/about.ts
export type AboutType = {
  section: string
  title: string
  content: string
  mapImage?: string
}

export type VisionMissionType = {
  vision: {
    title: string
    content: string
  }
  mission: {
    title: string
    content: string
  }
}

export type ProfileType = {
  heading: string
  img: string
  name: string
}